feat(analogical-reasoning): implement code mode architecture with layered structure
Introduce a three-layer architecture (Core, Code Mode, MCP) to separate concerns and improve maintainability. The Core layer contains domain logic and types, the Code Mode layer provides a public TypeScript API, and the MCP layer handles protocol adaptation. This refactoring maintains backward compatibility while enabling programmatic use of the analogical reasoning functionality.
